Knowledgebot/requirements.txt
fastapi
uvicorn[standard]
google-genai
scikit-learn
numpy
python-multipart

